@@ -4,7 +4,7 @@
     "minimum-stability": "stable",  
     "prefer-stable": true,
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "ext-ctype": "*",
         "ext-fileinfo": "*",
         "ext-gd": "*",
@@ -49,6 +49,7 @@
         "dama/doctrine-test-bundle": "7.1.1",
         "phpunit/php-code-coverage": "9.2.23",
         "phpunit/phpunit": "9.5.28",
+        "rector/rector": "0.15.7",
         "symfony/debug-bundle": "6.2.1",
         "symfony/maker-bundle": "1.48.0",
         "symfony/stopwatch": "6.2.0",
