@@ -10,6 +10,6 @@ if (file_exists(dirname(__DIR__).'/config/bootstrap.php')) {
     (new Dotenv())->bootEnv(dirname(__DIR__).'/.env');
 }
 
-if (false === (bool) $_SERVER['APP_DEBUG']) {
+if (!(bool) $_SERVER['APP_DEBUG']) {
     (new Symfony\Component\Filesystem\Filesystem())->remove(__DIR__.'/../var/cache/test');
 }
